services/photostock/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
boto3==1.39.8
python-dotenv==1.0.1
